made select_box one component
