+++ src/components/singleFunction/select_box.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import Select from 'react-select';
 
-export default class SelectSingle extends Component {
+export default class SelectMulti extends Component {
   constructor() {
     super();
     this.state = { selection: '' };
@@ -9,29 +9,33 @@ export default class SelectSingle extends Component {
     this.onInputChange = this.onInputChange.bind(this);
   }
 
-  onInputChange(value) {
-    console.log(value)
-    this.setState({ selection: value });
-    this.props.selection(value)
+  onInputChange(obj) {
+    console.log(obj)
+    this.setState({ selection: obj });
+    this.props.selection(obj)
   }
 
   render() {
     /*
     Important Notes:
 
+      This.props.multi = true or false;
+
       this.props.options = items within the select.
       They are an array of objects as follows:
       options = [{ value: -string/number-, label: -display string- }]
 
       OnChange will also return current selected result back to the calling component.
-      The return value is a number or a string value
+      The return value is an array of objects
       return prop = this.props.selection
     */
     const options = this.props.options;
+    const multi = this.props.multi;
     return (
       <div>
         <div>
           <Select
+            multi={ multi }
             name="form-field-select"
             className="Select"
             placeholder="make selection"
